Title: 최솟값 3개, Time: 176ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240503/최솟값 3개/top-3-smallest-number.cpp
@@ -13,7 +13,7 @@ int main()
         pq.push(-x);
         
         int a[3];
-        int mul = 1;
+        long mul = 1;
         if(i>=2)
         {
             for(int j=0;j<3;j++)
